+++ src/analyzer/tests/x-frame-options.js
@@ -1,8 +1,10 @@
 import { X_FRAME_OPTIONS } from "../../headers.js";
-import { BaseOutput, Requests } from "../../types.js";
-import { Expectation } from "../../types.js";
+import { BaseOutput, Expectation } from "../../types.js";
 import { getFirstHttpHeader } from "../utils.js";
+
 import { contentSecurityPolicyTest } from "./csp.js";
+
+/** @import { Requests } from "../../types.js" */
 
 export class XFrameOptionsOutput extends BaseOutput {
   /** @type {string | null} */
@@ -16,14 +18,6 @@ export class XFrameOptionsOutput extends BaseOutput {
     Expectation.XFrameOptionsNotImplemented,
     Expectation.XFrameOptionsHeaderInvalid,
   ];
-
-  /**
-   *
-   * @param {Expectation} expectation
-   */
-  constructor(expectation) {
-    super(expectation);
-  }
 }
 
 /**