+++ src/analyzer/utils.js
@@ -1,5 +1,6 @@
-import { parseItem, Token } from "structured-headers";
-import { Expectation } from "../types.js";
+import { Token, parseItem } from "structured-headers";
+
+/** @import { Expectation } from "../types.js" */
 
 /**
  * Return the new result if it's worse than the existing result, otherwise just the current result.
@@ -36,8 +37,7 @@ export function getHttpHeaders(response, name) {
     .filter(([headerName, _value]) => {
       return headerName.toLowerCase() === lcName;
     })
-    .map(([_headerName, value]) => value)
-    .flat();
+    .flatMap(([_headerName, value]) => value);
   return headers;
 }